+++ iop4admin/modeladmins/masterdark.py
@@ -50,11 +50,3 @@ class AdminMasterDark(AdminFitFile):
         url = reverse('iop4admin:%s_%s_changelist' % (MasterBias._meta.app_label, MasterBias._meta.model_name)) + f"?id={obj.masterbias.id}"
         return mark_safe(rf'<a href="{url}">{obj.masterbias.id}</a>')
     
-    @admin.display(description="Built from")
-    def get_built_from(self, obj):
-        self.allow_tags = True
-        
-        ids_str_L = [str(rf.id) for rf in obj.rawfits.all()]
-        a_href = reverse('iop4admin:%s_%s_changelist' % (RawFit._meta.app_label, RawFit._meta.model_name)) + "?id__in=%s" % ",".join(ids_str_L)
-        a_text = ", ".join(ids_str_L)
-        return mark_safe(f'<a href="{a_href}">{a_text}</a>')
